requirements.txt
pymongo==4.1.1
pymongo[srv]
python-dotenv==0.20.0
requests==2.27.1
selenium==4.3.0
tqdm==4.63.0
